pyproject.toml
[project]
name = "tp2-ai-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "pypdf>=6.9.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
